rnd-customer-segmentation-master/requirements.txt
absl-py==0.7.1
appdirs==1.4.3
asciimatics==1.11.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
cloudpickle==1.2.1
colorama==0.4.1
configobj==5.0.6
configparser==3.7.4
contextlib2==0.5.5
cycler==0.10.0
dask==2.1.0
decorator==4.4.0
defusedxml==0.6.0
distro==1.4.0
dvc==0.51.1
entrypoints==0.3
funcy==1.12
future==0.17.1
gast==0.2.2
gitdb2==2.0.5
GitPython==2.1.11
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-pasta==0.1.7
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grandalf==0.6
grpcio==1.21.1
h5py==2.9.0
humanize==0.5.1
idna==2.8
imageio==2.5.0
inflect==2.1.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.14.0
Jinja2==2.10.1
joblib==0.13.2
jsonpath-ng==1.4.3
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lightgbm==2.2.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
nanotime==0.5.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.3
notebook==5.7.8
numpy==1.16.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pathspec==0.5.9
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.2
ptyprocess==0.6.0
pyarrow==0.14.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyfiglet==0.8.post1
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
rsa==4.0
ruamel.yaml==0.15.98
schema==0.7.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
shap==0.29.3
shortuuid==0.5.0
six==1.12.0
smmap2==2.0.5
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019060501
toolz==0.9.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
treelib==1.5.5
Unidecode==1.1.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wordcloud==1.5.0
wrapt==1.11.2
xgboost==0.90
zc.lockfile==1.4
